requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
# requirements.txt
reportlab>=3.6.12

pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
